adopted better naming conventions for post routes
Used id and publicId params to both refer to public Ids available for resources. In cases where parent resource only is involved, publicId is used as params and in cases where child recources are involved like comments, id is used as param for parent resource and publicId is used as param for child reource. Also whenever mongo Ids are involved, name of reource is prepended to Id e.g postId and commentId